+++ HISTORY.rst
@@ -1,6 +1,12 @@
 History
 =======
 
+* Add defaults for ``SERVER_HOST``, ``SERVER_PORT`` and ``wsgi.url_scheme``.
+  This enables responding to `ELB health check events
+  <https://docs.aws.amazon.com/elasticloadbalancing/latest/application/lambda-functions.html#enable-health-checks-lambda>`__,
+  which don't contain the relevant headers
+  (`Issue #155 <https://github.com/adamchainz/apig-wsgi/pull/155>`__).
+
 2.6.0 (2020-03-07)
 ------------------